Renaming attachmentService
RISDEV-3709
